Code cleanup
Use diff -U 3 when running the testsuite, for better feedback.

Use size_t instead of int, for parameter and callback types, where
appropriate.

Pass an entire string chunk to functions and callbacks, instead of a
single character.

+++ rfc2045/rfc2045decodemsgtoutf8.c
@@ -53,7 +53,7 @@ static void doconvtoutf8_write_noeol(const char *p, size_t n,
 	}
 }
 
-static void doconvtoutf8_error(const char *p, int n,
+static void doconvtoutf8_error(const char *p, size_t n,
 			       void *void_arg)
 {
 	struct doconvtoutf8_info *ptr=(struct doconvtoutf8_info *)void_arg;
